+++ NEWS.md
@@ -1,3 +1,9 @@
+# confintr 0.1.1
+
+- Added confidence intervals for the odds ratio.
+
+- Fixed a wrong "VignetteIndexEntry".
+
 # confintr 0.1.0
 
 This is the initial CRAN release.